feat(predict): cover any-token withdraw in Appium smoke
Drop the Detox SmokePredictions restore and move CONF-918 coverage next to
the existing Appium predict withdraw suite so CI runs without workflow changes.
