+++ .github/workflows/run-e2e-smoke-tests-android.yml
@@ -58,23 +58,6 @@ jobs:
       runner_provider: ${{ inputs.runner_provider }}
     secrets: inherit
 
-  prediction-market-android-smoke:
-    if: contains(fromJson(inputs.selected_tags), 'ALL') || contains(fromJson(inputs.selected_tags), 'SmokePredictions')
-    strategy:
-      matrix:
-        split: [1]
-      fail-fast: false
-    uses: ./.github/workflows/run-e2e-workflow.yml
-    with:
-      test-suite-name: prediction-market-android-smoke-${{ matrix.split }}
-      platform: android
-      test_suite_tag: 'SmokePredictions'
-      split_number: ${{ matrix.split }}
-      total_splits: 1
-      changed_files: ${{ inputs.changed_files }}
-      runner_provider: ${{ inputs.runner_provider }}
-    secrets: inherit
-
   wallet-platform-android-smoke:
     if: contains(fromJson(inputs.selected_tags), 'ALL') || contains(fromJson(inputs.selected_tags), 'SmokeWalletPlatform')
     strategy:
@@ -210,7 +193,6 @@ jobs:
     needs:
       - stake-android-smoke
       - perps-android-smoke
-      - prediction-market-android-smoke
       - wallet-platform-android-smoke
       - network-abstraction-android-smoke
       - network-expansion-android-smoke
